requirements.txt
streamlit
Pillow
SQLAlchemy
psycopg2-binary